+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adding image size properties to the infoCard component
+
 ## [3.176.1] - 2024-11-08
 
 ## [3.176.0] - 2024-11-06